.NET 6 + random id's for triggers and variables

+++ BetterTriggers/Controllers/ControllerVariable.cs
@@ -159,22 +159,22 @@ namespace BetterTriggers.Controllers
             return list;
         }
 
-        public static string GetVariableNameById(int id)
+        public static string GetVariableNameById(long id)
         {
             return Variables.GetVariableNameById(id);
         }
 
-        public static Variable GetById(int id)
+        public static Variable GetById(long id)
         {
             return Variables.GetVariableById(id);
         }
 
-        public static Variable GetById(int id, Trigger trig)
+        public static Variable GetById(long id, Trigger trig)
         {
             return Variables.GetVariableById(id, trig);
         }
 
-        public static Variable GetById_AllLocals(int id)
+        public static Variable GetById_AllLocals(long id)
         {
             return Variables.GetVariableById_AllLocals(id);
         }
