+++ BetterTriggers/Models/EditorData/ExplorerElementScript.cs
@@ -87,7 +87,7 @@ namespace BetterTriggers.Models.EditorData
             }
         }
 
-        public int GetId()
+        public long GetId()
         {
             throw new NotImplementedException();
         }
