+++ BetterTriggers/Models/SaveableData/TriggerRef.cs
@@ -1,5 +1,6 @@
 ﻿using BetterTriggers.JsonBaseConverter;
 using Newtonsoft.Json;
+using System;
 using System.Collections.Generic;
 
 namespace BetterTriggers.Models.SaveableData
@@ -7,7 +8,7 @@ namespace BetterTriggers.Models.SaveableData
     public class TriggerRef : Parameter
     {
         public readonly int ParamType = 4; // DO NOT CHANGE
-        public int TriggerId;
+        public long TriggerId;
 
         public override TriggerRef Clone()
         {
